requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-multiselectfield==0.1.12
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pylint==2.6.0
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2020.4
razorpay==1.2.0
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
